Try to prove that keepAlive is to blame.
The problem also doesn't exist on Node 0.10 (first shows in 0.12, which changed something about the way `Keep-Alive` worked.

If I create my own `require('https').Agent` with `keepAlive: true`, then the problem goes away. If I instead set `keepAlive: false` on that agent - it fails.  If I don't pass my own agent (and let it default to whatever it does) it fails.

If I change the Network implementation of my VirtualBox to "Bridged Adapter" instead of NAT, the problem goes away.

This finally proved to me that it was a networking issue.  Interestingly, setting `keepAlive: true` fixes it, but I'm not sure what the implications of doing that are.  This leads me to this bug:

https://www.virtualbox.org/ticket/14049

Joy? Maybe. I guess don't use "NAT" networking in VirtualBox.